split coordinator register pages up made submit buttons more similair
